# swagger.json
{"openapi":"3.0.1","info":{"title":"Cadastro de Beneficiario Demo","description":"Aplicação Restfull para cadastro de beneficiario.\r\nPara acessar as operação utilisar autenticação básica com o usuario \"user\" e senha  12345","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/v1/beneficiarios/{id}":{"get":{"tags":["beneficiario-controller"],"description":"Busca o beneficiarios utilizando o Id","operationId":"buscaPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Retorna o beneficiario","content":{"application/json":{}}},"500":{"description":"Erro na busca do beneficiario","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorVo"}}}}}},"put":{"tags":["beneficiario-controller"],"description":"Atualiza o beneficiário passando o id","operationId":"atualizaBeneficiario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CriaBeneficiarioDTO"}}},"required":true},"responses":{"500":{"description":"Erro na atualização do beneficiario","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorVo"}}}},"200":{"description":"Sucesso","content":{"application/json":{}}}}},"delete":{"tags":["beneficiario-controller"],"description":"Remove o beneficiário pelo id","operationId":"removerBeneficiario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"500":{"description":"Erro na remocao do beneficiario","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorVo"}}}},"200":{"description":"Sucesso","content":{"application/json":{}}}}}},"/v1/beneficiarios":{"get":{"tags":["beneficiario-controller"],"description":"Busca todos os beneficiarios","operationId":"buscaBeneficiarios","responses":{"200":{"description":"Retorna os beneficiarios","content":{"application/json":{}}},"500":{"description":"Erro na busca dos beneficiarios","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorVo"}}}}}},"post":{"tags":["beneficiario-controller"],"description":"Cria um novo beneficiario com os documentos","operationId":"criarBeneficiario","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CriaBeneficiarioDTO"}}},"required":true},"responses":{"500":{"description":"Erro na criação do beneficiario","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorVo"}}}},"200":{"description":"Retorna o beneficiario criado com o id","content":{"application/json":{}}}}}},"/v1/gretting":{"get":{"tags":["beneficiario-controller"],"operationId":"greeting","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/v1/beneficiarios/{id}/documentos":{"get":{"tags":["beneficiario-controller"],"description":"Busca os documentos do beneficiario","operationId":"buscaDocumentos","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Retorna os documentos","content":{"application/json":{}}},"500":{"description":"Erro na busca dos documentos","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorVo"}}}}}}}},"components":{"schemas":{"CriaBeneficiarioDTO":{"required":["dataNascimento","documentos","nome","telefone"],"type":"object","properties":{"nome":{"type":"string"},"telefone":{"type":"string"},"dataNascimento":{"type":"string"},"documentos":{"type":"array","items":{"$ref":"#/components/schemas/DocumentoDTO"}}}},"DocumentoDTO":{"required":["dadosDocumento","descricao","tipoDocumento"],"type":"object","properties":{"tipoDocumento":{"type":"string"},"descricao":{"type":"string"},"dadosDocumento":{"type":"string"}}},"ErrorVo":{"type":"object","properties":{"message":{"type":"string"},"url":{"type":"string"},"stackTrace":{"type":"string"}}}}}}